Change sincBlackmanInterp2.m to return single.  Add DX,DY,A return from subpixel.m for LSQ and 3pG

+++ sincBlackmanInterp2.m
@@ -180,7 +180,8 @@ clear sourceRows sourceColumns
 % evaluated at each pixel. Doing this as an array operation is faster than
 % doing it in a loop because it takes a long time to loop over a million
 % goddamn pixels. 
-Zinterp = uint8( sum( Weights .* single(sourcePixels) , 2 ) );
+% Zinterp = uint8( sum( Weights .* single(sourcePixels) , 2 ) );
+Zinterp = single( sum( Weights .* single(sourcePixels) , 2 ) );
 
 % Clear a variable to save space
 clear sourcePixels
